+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "bcrypt-nodejs": "0.0.3",
     "body-parser": "^1.18.3",
+    "clarifai": "^2.8.1",
     "cors": "^2.8.4",
     "express": "^4.16.3",
     "knex": "^0.15.2",
